[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -7,8 +7,8 @@ int main() {
     cin.tie(NULL);
 
     int L;
-    int hash = 0;
-    int R = 1;
+    long long hash = 0;
+    long long R = 1;
     int r = 31;
     int M = 1234567891;
     char arr[51];
